requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
sentence-transformers==3.4.1
faiss-cpu==1.7.3
httpx==0.24.1
pytest==7.4.0
